requirements.txt
python-telegram-bot==20.3
pymongo
google-generativeai
python-dotenv
requests